requirements.txt
Django~=4.1.3
python-dotenv~=0.21.0
pytz~=2022.6
PyJWT~=2.6.0
djangorestframework~=3.14.0
zeep~=4.2.0
